% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmospheric_attenuation.R
\name{atmospheric_attenuation}
\alias{atmospheric_attenuation}
\title{Measure atmospheric attenuation and absorption of sound}
\usage{
atmospheric_attenuation(f, temp, RH, p = 101325, 
formula = 1, spi = NULL, dist = NULL)
}
\arguments{
\item{f}{numeric vector of length 1 with frequency (in Hertz).}

\item{temp}{numeric vector of length 1 with frequency (in Celsius).}

\item{RH}{numeric vector of length 1 with relative humidity}

\item{p}{numeric vector of length 1 with ambient pressure in Pa (standard: 101325, default).}

\item{formula}{1 = Bazley 1976 (used by Marc), 2 = Rossing 2007 (p. 116, see details).}

\item{spi}{numeric vector of length 1 with the initial sound pressure in Pa. Required for calculating atmospheric absorption. Default is \code{NULL}.}

\item{dist}{numeric vector of length 1 with the distance in m through which the sound propagates. Required for calculating atmospheric absorption. Default is \code{NULL}.}
}
\value{
Returns the atmospheric absorption (in dB/m) of sound based on the supplied parameters. I
}
\description{
\code{atmospheric_attenuation} measures atmospheric attenuation and atmospheric absorption of signals referenced in a extended selection table.
}
\details{
Calculate the atmospheric attenuation in dependence on temperature, relative humidity, pressure and frequency. The function can applied to formulae based on:
\itemize{
\item \code{1}: default. As used by Marc: Bazley (1976), Sound absorption in air at
\item \code{2}: as used by Peter: Rossing (2007), Handbook of Acoustics, Springer.
}
If 'spi' is supplied the function also returns the atmospheric absorption (in dB).
}
\examples{
{
# load example data
data("playback_est")

#' # remove noise selections
playback_est <- playback_est[playback_est$signal.id != "noise", ]

# measure atmospheric attenuation formula 1
atmospheric_attenuation(f = 20000, temp = 20, RH = 90, p = 88000, formula = 1)
}

}
\author{
Marcelo Araya-Salas (\email{marceloa27@gmail.com}) 
#' @references {
Araya-Salas, M. (2019), baRulho: a R package to evaluate habitat-induced degradation of (animal) acoustic signals. R package version 1.0.0
}
}
