% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baRulho-package.R
\docType{package}
\name{baRulho}
\alias{baRulho}
\alias{baRulho-package}
\title{baRulho: quantifying habitat-induced acoustic signal degradation}
\description{
`baRulho` is a package intended to quantify habitat-induced degradation of (animal) acoustic signals.
}
\details{
The main features of the package are:
  \itemize{
  \item Loops to apply tasks through sounds referenced in an extended selection table
  \item The comparison of playback sounds re-recorded at different distances
  }
Most functions allow the parallelization of tasks, which distributes the tasks among several processors to improve computational efficiency.

License: GPL (>= 2)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maRce10/baRulho}
  \item Report bugs at \url{https://github.com/maRce10/baRulho/issues}
}

}
\author{
Marcelo Araya-Salas

  Maintainer: Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
