% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degradation_est-data.R
\docType{data}
\name{degradation_est}
\alias{degradation_est}
\title{Extended selection table with re-recorded playbacks}
\format{
Extended selection table object in the \code{\link{warbleR}} format, which contains annotations and acoustic data.
}
\source{
Marcelo Araya-Salas
}
\usage{
data(degradation_est)
}
\description{
Master and re-recorded 2 s long, frequency modulated, amplitude modulated, synthetic sounds that were broadcast and re-recorded at 3 distances (1, 10 and 30m, column 'distance'). The data was created by the function \code{\link[warbleR]{selection_table}} from the warbleR package.
}
\keyword{datasets}
