% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_distance.R
\name{detection_distance}
\alias{detection_distance}
\title{Measure detection distance of sound}
\usage{
detection_distance(
  X,
  cores = getOption("mc.cores", 1),
  pb = getOption("pb", TRUE),
  hop.size = getOption("hop.size", 11.6),
  wl = getOption("wl", NULL),
  path = getOption("sound.files.path", "."),
  spl = NULL,
  spl.cutoff = NULL,
  temp = 20,
  rh = 60,
  pa = 101325,
  hab.att.coef = 0.02,
  max.distance = 1000,
  resolution = 0.1,
  subtract.bgn = TRUE,
  envelope = c("abs", "hil"),
  mar = NULL
)
}
\arguments{
\item{X}{The output of \code{\link[=set_reference_sounds]{set_reference_sounds()}}, an object of class
\code{data.frame}, \code{selection_table}, or \code{extended_selection_table} (the
last 2 classes are created by \code{\link[warbleR:selection_table]{warbleR::selection_table()}} from the
\strong{warbleR} package) with the test sound files' annotations. Must
contain the following columns: 1) \code{sound.files}: name of the \code{.wav}
files, 2) \code{selec}: unique selection identifier (within a sound
file), 3) \code{start}: start time and 4) \code{end}: end time of selections,
5) \code{bottom.freq}: low frequency for bandpass, 6) \code{top.freq}: high
frequency for bandpass, 7) \code{sound.id}: ID of sounds used to
identify counterparts across distances, and 8) \code{reference}:
identity of sounds to be used as reference for each test sound
(row). See \code{\link[=set_reference_sounds]{set_reference_sounds()}} for more details on the
structure of \code{X}.}

\item{cores}{Numeric vector of length 1. Controls whether parallel
computing is applied by specifying the number of cores to be used.
Default \code{1} (i.e. no parallel computing). Can be set globally for
the current R session via the \code{"mc.cores"} option (see
\code{\link[=options]{options()}}).}

\item{pb}{Logical argument to control if progress bar is shown.
Default \code{TRUE}. Can be set globally for the current R session via
the \code{"pb"} option (see \code{\link[=options]{options()}}).}

\item{hop.size}{Numeric vector of length 1 specifying the time window
duration (in ms). Default \code{11.6} ms, which is equivalent to 512
\code{wl} for a 44.1 kHz sampling rate. Ignored if \code{wl} is supplied. Can
be set globally for the current R session via the \code{"hop.size"}
option (see \code{\link[=options]{options()}}).}

\item{wl}{A vector with a single even integer number specifying the
window length of the spectrogram. Default \code{NULL}. If supplied,
\code{hop.size} is ignored. Odd integers will be rounded up to the
nearest even number. Can be set globally for the current R session
via the \code{"wl"} option (see \code{\link[=options]{options()}}).}

\item{path}{Character string containing the directory path where the
sound files are found. Only needed when \code{X} is not an extended
selection table. If not supplied the current working directory is
used. Can be set globally for the current R session via the
\code{"sound.files.path"} option (see \code{\link[=options]{options()}}).}

\item{spl}{Numeric vector of length 1 specifying the sound pressure
level of sounds. If not supplied, it will be measured from the
sounds themselves.}

\item{spl.cutoff}{Numeric vector of length 1 specifying the sound
pressure level cutoff to define if the sound is no longer
detected. Ideally it should be estimated based on the sound
detection threshold of the species.}

\item{temp}{Numeric vector of length 1 with temperature (in
Celsius). Default \code{20}.}

\item{rh}{Numeric vector of length 1 with relative humidity (in
percentage). Default \code{60}.}

\item{pa}{Numeric vector of length 1 with ambient pressure in Pa
(standard: \code{101325}, default). Used for atmospheric attenuation.}

\item{hab.att.coef}{Attenuation coefficient of the habitat (in
dB/kHz/m).}

\item{max.distance}{Numeric vector of length 1 with the maximum
distance (in m) at which detection would be evaluated. Note that
the function calculates the expected sound pressure level values
along a vector of distances to find the distance at which the
expected sound pressure level equates \code{spl.cutoff}. Default
\code{1000} m.}

\item{resolution}{Numeric vector of length 1 with the distance
resolution (in m) for estimated detection distance. Higher
resolutions take longer to estimate. Default \code{0.1} m.}

\item{subtract.bgn}{Logical argument to control if SPL from
background noise is excluded from the measured signal SPL. Default
\code{FALSE}.}

\item{envelope}{Character string vector with the method to calculate
amplitude envelopes (in which SPL is measured, only used if \code{spl}
is not supplied), as in \code{\link[seewave:env]{seewave::env()}}. Must be either \code{"abs"}
(absolute envelope, default) or \code{"hil"} (Hilbert transformation).}

\item{mar}{Numeric vector of length 1. Specifies the margins
adjacent to the start and end points of selection over which to
measure background noise. This is required to subtract background
noise sound pressure level (so only needed when
\code{subtract.bgn = TRUE}).}
}
\value{
Object \code{X} with an additional column, \code{detection.distance},
containing the computed detection distances (in m).
}
\description{
\code{detection_distance()} estimates the detection distance of sounds
referenced in an extended selection table.
}
\details{
The function computes the maximum distance at which a sound would be
detected, which is calculated as the distance at which the sound
pressure level (SPL) goes below the specified SPL cutoff
(\code{spl.cutoff}). This is returned as an additional column,
\code{detection.distance} (in m). The function uses \code{\link[=attenuation]{attenuation()}}
internally to estimate SPL at increasing distances until it reaches
the defined cutoff. The peak frequency (calculated on the power
spectrum of the reference sound) of the reference sound for each
sound ID is used as the carrier frequency for distance estimation.
The sound recorded at the lowest distance is used as reference.
\strong{This function assumes that all recordings have been made at the
same recording volume.}
}
\examples{
\dontrun{
# load example data
data("test_sounds_est")

# add reference to X
X <- set_reference_sounds(X = test_sounds_est)

detection_distance(X = X[X$distance \%in\% c(1, 10), ], spl.cutoff = 5, mar = 0.05)
}

}
\references{
Araya-Salas, M., Grabarczyk, E. E., Quiroz-Oliva, M., Garcia-Rodriguez, A., & Rico-Guevara, A. (2025). Quantifying degradation in animal acoustic signals with the R package baRulho. Methods in Ecology and Evolution, 00, 1-12. https://doi.org/10.1111/2041-210X.14481

Clark, C.W., Marler, P. & Beeman K. (1987). Quantitative analysis of animal vocal phonology: an application to Swamp Sparrow song. Ethology. 76:101-115.
}
\seealso{
\code{\link[=attenuation]{attenuation()}}, used internally to estimate SPL at
increasing distances.

Other quantify degradation: 
\code{\link{blur_ratio}()},
\code{\link{envelope_correlation}()},
\code{\link{plot_blur_ratio}()},
\code{\link{plot_degradation}()},
\code{\link{set_reference_sounds}()},
\code{\link{signal_to_noise_ratio}()},
\code{\link{spcc}()},
\code{\link{spectrum_blur_ratio}()},
\code{\link{spectrum_correlation}()},
\code{\link{tail_to_signal_ratio}()}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
\concept{quantify degradation}
