% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_att.R
\name{excess_att}
\alias{excess_att}
\title{Measure excess attenuation}
\usage{
excess_att(X, mar, parallel = 1, pb = TRUE, method = 1, eq.dur = FALSE,
noise.ref = "before", bp = NULL, wl = 10, output = "est")
}
\arguments{
\item{X}{object of class 'selection_table', 'extended_selection_table' created by the function \code{\link[warbleR]{selection_table}} from the warbleR package.}

\item{mar}{numeric vector of length 1. Specifies the margins adjacent to
the start and end points of selection over which to measure noise.}

\item{parallel}{Numeric. Controls whether parallel computing is applied.
It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).
If \code{NULL} (default) then the current working directory is used.}

\item{pb}{Logical argument to control if progress bar is shown. Default is \code{TRUE}. It can also be
set globally using the 'pb' option (see \code{\link{warbleR_options}}).}

\item{method}{Numeric vector of length 1 to indicate the 'experimental design' for measuring excess attenuation. Two methods are available:
\itemize{
\item \code{1}: compare all signals with their counterpart that was recorded at the closest distance to source (e.g. compare a signal recorded at 5m, 10m and 15m with its counterpart recorded at 1m). This is the default method. 
\item \code{2}: compare all signals with their counterpart recorded at the distance inmediately before (e.g. a signal recorded at 10m compared with the same signal recorded at 5m, then signal recorded at 15m compared with same signal recorded at 10m and so on).
}}

\item{eq.dur}{Logical. Controls whether the noise segment that is measured has the same duration 
than the signal (if \code{TRUE}, default \code{FALSE}). If \code{TRUE} then 'mar' argument is ignored.}

\item{noise.ref}{Character with the name of the column that contains tags for "signal" and "noise". This can be used to apply the same noise reference to all signals in a sound file. Therefore, at least one "noise" selection for each sound file must be provided. Default is \code{NULL}.}

\item{bp}{Numeric vector of length 2 giving the lower and upper limits of a frequency bandpass filter (in kHz). Default is \code{NULL}.}

\item{wl}{A numeric vector of length 1 specifying the window length of the spectrogram for applying bandpass. Default 
 is 10. Ignored if \code{bp = NULL}.
Note that lower values will increase time resolution, which is more important for amplitude ratio calculations.}

\item{output}{Character vector of length 1 to determine if an extended selection table ('est', default) or a data frame ('data.frame') is returned.}
}
\value{
Data frame or extended selection table (depending on 'output' argument) similar to input data, but also includes a new column 
with the excess attenuation values.
}
\description{
\code{excess_att} Measure excess attenuation in signals referenced in a extended selection table.
}
\details{
Excess attenuation is the attenuation of a sound in excess of that due to spherical spreading as described by Dabelsteen et al (1993). The goal of the function is to measure the excess attenuation on signals in which a master playback has been re-recorded at different distances. The 'signal.id' column must be used to indicate which signals belonging to the same category (e.g. song-types). The function will then compared each signal type to its reference. Two methods for calculating excess attenuation are provided (see 'method' argument).
}
\examples{
{
# First set temporary folder
# setwd(tempdir())

data("playback_est")

# using margin for noise and method 1
excess_att(X = playback_est, mar = 0.05)

# using margin for noise and method 2
excess_att(X = playback_est, mar = 0.05, method = 2)
}

}
\author{
Marcelo Araya-Salas (\email{marceloa27@gmail.com}) #' @references {
Dabelsteen, T., Larsen, O. N., & Pedersen, S. B. (1993). Habitat-induced degradation of sound signals: Quantifying the effects of communication sounds and bird location on blur ratio, excess attenuation, and signal-to-noise ratio in blackbird song. The Journal of the Acoustical Society of America, 93(4), 2206.
}
}
