% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers.R
\name{find_markers}
\alias{find_markers}
\title{Find acoustic markers on test sound files}
\usage{
find_markers(
  X,
  markers = c("start_marker", "end_marker"),
  test.files = NULL,
  path = getOption("sound.files.path", "."),
  pb = getOption("pb", TRUE),
  cores = getOption("mc.cores", 1),
  ...
)
}
\arguments{
\item{X}{Object of class \code{data.frame}, \code{selection_table}, or
\code{extended_selection_table} (the last 2 classes are created by
\code{\link[warbleR:selection_table]{warbleR::selection_table()}} from the \strong{warbleR} package) with
the reference to the sounds in the master sound file. Must
contain the following columns: 1) \code{sound.files}: name of the
\code{.wav} files, 2) \code{selec}: unique selection identifier (within a
sound file), 3) \code{start}: start time, 4) \code{end}: end time of
selections, and 5) \code{sound.id}: unique identifier for each of the
annotated sounds in \code{X}. Columns for \code{top.freq}, \code{bottom.freq},
and \code{channel} are optional. The acoustic start and end markers
(added by \code{\link[=master_sound_file]{master_sound_file()}}) should be labeled as
\code{"start_marker"} and \code{"end_marker"} respectively. Required.}

\item{markers}{Character vector with the name of the annotations (as
in the column \code{sound.id}) to be used as templates for
cross-correlation. Default \code{c("start_marker", "end_marker")}.
Using more than one marker is recommended, as the time difference
between their position can be used to evaluate the precision of
the detection (see the Value section).}

\item{test.files}{Character vector of length 1 with the name(s) of
the test (re-recorded) file(s) in which to search for the
marker(s). If not supplied, all sound files in \code{path} are used
instead.}

\item{path}{Character string containing the directory path where
test (re-recorded) sound files are found.}

\item{pb}{Logical argument to control if progress bar is shown.
Default \code{TRUE}. Can be set globally for the current R session via
the \code{"pb"} option (see \code{\link[=options]{options()}}).}

\item{cores}{Numeric vector of length 1. Controls whether parallel
computing is applied by specifying the number of cores to be used.
Default \code{1} (i.e. no parallel computing). Can be set globally for
the current R session via the \code{"mc.cores"} option (see
\code{\link[=options]{options()}}).}

\item{...}{Additional arguments to be passed to
\code{\link[ohun:template_correlator]{ohun::template_correlator()}} for setting cross-correlation
parameters (e.g. \code{wl}, \code{ovlp}, etc).}
}
\value{
A \code{data.frame} with test file names, marker ID, maximum
cross-correlation score for each marker, and the start and end where
it was detected. If two or more markers are used, the function
computes an additional column, \code{time.mismatch}, that compares the
time difference between the two markers in the test files against
that in the master sound file. In a perfect detection, the value must
be 0.
}
\description{
\code{find_markers()} finds acoustic markers on test (re-recorded) sound
files using spectrographic cross-correlation.
}
\details{
The function takes a master sound file's reference data (\code{X}) and
finds the position of acoustic markers (\code{markers} argument, included
as selections in \code{X}) in the re-recorded sound files. This is used
to align signals found in re-recorded sound files according to a
master sound file referenced in \code{X}. The position of the markers is
determined as the highest spectrogram cross-correlation value for
each marker using the functions \code{\link[ohun:template_correlator]{ohun::template_correlator()}} and
\code{\link[ohun:template_detector]{ohun::template_detector()}}. \strong{Make sure the master sound file
(referred to in \code{X}) is found in the same folder as the re-recorded
sound files.} Take a look at the package vignette for information
on how to incorporate this function into a sound degradation
analysis workflow.

In cases in which markers are not correctly detected, editing test
sound files to remove audio segments with no target sounds (before
the start marker and after the end marker) can improve performance.
Using a low \code{hop.size} or window length \code{wl} (used internally by
\code{\link[ohun:template_correlator]{ohun::template_correlator()}}) can help to improve precision. Other
spectrogram types (argument \code{type} in \code{\link[ohun:template_correlator]{ohun::template_correlator()}})
can sometimes show better performance when markers are highly
degraded. If frequency range columns are included (\code{bottom.freq} and
\code{top.freq}, in kHz), cross-correlation will be run on those frequency
ranges. All templates must have the same sampling rate, and both
templates and \code{files} (in which to find templates) must also have
the same sampling rate.
}
\examples{
{
  # set temporary directory
  td <- tempdir()

  # load example data
  data("master_est")

  # save example files in working director to recreate a case in which working
  # with sound files instead of extended selection tables.
  # This doesn't have to be done with your own data as you will
  # have them as sound files already.
  for (i in unique(test_sounds_est$sound.files)[1:2]) {
    writeWave(object = attr(test_sounds_est, "wave.objects")[[i]], file.path(td, i))
  }

  # save master file
  writeWave(object = attr(master_est, "wave.objects")[[1]], file.path(td, "master.wav"))

  # set path and no progress bar in global options
  options(sound.files.path = td, pb = FALSE)

  # get marker position
  markers <- find_markers(X = master_est, test.files = unique(test_sounds_est$sound.files)[2])
}
}
\references{
Araya-Salas, M., Grabarczyk, E. E., Quiroz-Oliva, M., Garcia-Rodriguez, A., & Rico-Guevara, A. (2025). Quantifying degradation in animal acoustic signals with the R package baRulho. Methods in Ecology and Evolution, 00, 1-12. https://doi.org/10.1111/2041-210X.14481
}
\seealso{
\code{\link[=manual_realign]{manual_realign()}}, \code{\link[=auto_realign]{auto_realign()}}, and
\code{\link[=align_test_files]{align_test_files()}}, which align the re-recorded sounds once
markers have been found; \code{\link[=master_sound_file]{master_sound_file()}}, which creates the
markers in the first place.

Other test sound alignment: 
\code{\link{align_test_files}()},
\code{\link{auto_realign}()},
\code{\link{manual_realign}()},
\code{\link{plot_aligned_sounds}()}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
\concept{test sound alignment}
