% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_profile.R
\name{noise_profile}
\alias{noise_profile}
\title{Measure full spectrum sound noise profiles}
\usage{
noise_profile(
  X = NULL,
  files = NULL,
  mar = NULL,
  noise.ref = c("adjacent", "custom"),
  cores = getOption("mc.cores", 1),
  pb = getOption("pb", TRUE),
  path = getOption("sound.files.path", "."),
  bp = NULL,
  hop.size = getOption("hop.size", 1),
  wl = getOption("wl", NULL),
  PSD = FALSE,
  norm = TRUE,
  dB = c("A", "B", "C", "D", "ITU", "max0"),
  averaged = TRUE
)
}
\arguments{
\item{X}{Object of class \code{data.frame}, \code{selection_table}, or
\code{extended_selection_table} (the last 2 classes are created by
\code{\link[warbleR:selection_table]{warbleR::selection_table()}} from the \strong{warbleR} package) with
the test sound files' annotations. Must contain the following
columns: 1) \code{sound.files}: name of the \code{.wav} files, 2) \code{selec}:
unique selection identifier (within a sound file), 3) \code{start}:
start time and 4) \code{end}: end time of selections, 5)
\code{bottom.freq}: low frequency for bandpass, 6) \code{top.freq}: high
frequency for bandpass, and 7) \code{sound.id}: ID of sounds used to
identify counterparts across distances (needed for \code{"custom"}
noise reference, see the \code{noise.ref} argument). Default \code{NULL}.}

\item{files}{Character vector with names of wave files to be
analyzed. Files must be found in the supplied \code{path} (or in the
working directory if \code{path} is not supplied). Default \code{NULL}.}

\item{mar}{Numeric vector of length 1. Specifies the margins
adjacent to the start and end points of a selection over which to
measure ambient noise. Required if \code{X} is supplied, and ignored
if \code{X} is not supplied. Default \code{NULL}.}

\item{noise.ref}{Character vector of length 1 determining which
noise segment must be used for measuring ambient noise. Ignored
if \code{X} is not supplied. Two options are available:
\itemize{
\item \strong{\code{adjacent}}: measure ambient noise right before the sound
(using the \code{mar} argument to define duration of ambient noise
segments).
\item \strong{\code{custom}}: measure ambient noise segments referenced in the
selection table (labeled as \code{"ambient"} in the \code{sound.id}
column).
}}

\item{cores}{Numeric vector of length 1. Controls whether parallel
computing is applied by specifying the number of cores to be used.
Default \code{1} (i.e. no parallel computing). Can be set globally for
the current R session via the \code{"mc.cores"} option (see
\code{\link[=options]{options()}}).}

\item{pb}{Logical argument to control if progress bar is shown.
Default \code{TRUE}. Can be set globally for the current R session via
the \code{"pb"} option (see \code{\link[=options]{options()}}).}

\item{path}{Character string containing the directory path where the
sound files are found. Only needed when \code{X} is not an extended
selection table. If not supplied the current working directory is
used. Can be set globally for the current R session via the
\code{"sound.files.path"} option (see \code{\link[=options]{options()}}).}

\item{bp}{Numeric vector of length 2 giving the lower and upper
limits of a frequency bandpass filter (in kHz). Default \code{NULL}.}

\item{hop.size}{Numeric vector of length 1 specifying the time
window duration (in ms). Default \code{1} ms, which is equivalent to
~45 \code{wl} for a 44.1 kHz sampling rate. Ignored if \code{wl} is
supplied. Can be set globally for the current R session via the
\code{"hop.size"} option (see \code{\link[=options]{options()}}).}

\item{wl}{Numeric vector of length 1 specifying the window length
of the spectrogram. Default \code{NULL}. Ignored if \code{bp = NULL}. If
supplied, \code{hop.size} is ignored. Note that lower values will
increase time resolution, which is more important for amplitude
ratio calculations.}

\item{PSD}{Logical to control whether the Probability Mass Function
(the probability distribution of frequencies) is returned. See
\code{\link[seewave:meanspec]{seewave::meanspec()}}. Default \code{FALSE}.}

\item{norm}{Logical to control whether amplitude values are
normalized (divided by the maximum) so the highest value is 1.
See \code{\link[seewave:meanspec]{seewave::meanspec()}}. Default \code{TRUE}.}

\item{dB}{Character string of length 1 specifying the type of dB to
return: \code{"max0"} for a maximum dB value at 0, \code{"A"}, \code{"B"}, \code{"C"},
\code{"D"}, and \code{"ITU"} for common dB weights. See
\code{\link[seewave:meanspec]{seewave::meanspec()}}. Default \code{"A"}.}

\item{averaged}{Logical to control if frequency spectra are
averaged within a sound file. Default \code{TRUE}.}
}
\value{
A \code{data.frame} containing the frequency spectra for each sound file
or wave object (if \code{X} is supplied and is of class
\code{extended_selection_table}).
}
\description{
\code{noise_profile()} measures full spectrum sound pressure levels (i.e.
noise profiles) in sound files or extended selection tables.
}
\details{
The function estimates full spectrum sound pressure levels (i.e.
noise profiles) of ambient noise. This can be done on data
frames/(extended) selection tables (using the segments containing no
target sound or the \code{"ambient"} sound ID) or over complete sound
files in the working directory (or supplied \code{path}). The function
uses \code{\link[seewave:meanspec]{seewave::meanspec()}} internally to calculate frequency
spectra.
}
\examples{
{
  # load example data
  data("test_sounds_est")

  # measure on custom noise reference
  noise_profile(X = test_sounds_est, mar = 0.01, pb = FALSE, noise.ref = "custom")

  # remove noise selections so noise is measured right before the signals
  pe <- test_sounds_est[test_sounds_est$sound.id != "ambient", ]

  noise_profile(X = pe, mar = 0.01, pb = FALSE, noise.ref = "adjacent")
}
}
\references{
Araya-Salas, M., Grabarczyk, E. E., Quiroz-Oliva, M., Garcia-Rodriguez, A., & Rico-Guevara, A. (2025). Quantifying degradation in animal acoustic signals with the R package baRulho. Methods in Ecology and Evolution, 00, 1-12. https://doi.org/10.1111/2041-210X.14481.
}
\seealso{
\code{\link{excess_attenuation}}

Other miscellaneous: 
\code{\link{add_noise}()},
\code{\link{attenuation}()}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
\concept{miscellaneous}
