% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playback_est-data.R
\docType{data}
\name{playback_est}
\alias{playback_est}
\title{Extended selection table with re-recorded playbacks}
\format{
Extended selection table object in the \code{\link{warbleR}} format, which contains annotations and acoustic data
}
\source{
Marcelo Araya-Salas
}
\usage{
data(playback_est)
}
\description{
Recordings of \emph{Phaethornis longirostris} (Long-billed Hermit) songs from different song types (column 'signal.type') that were broadcast and re-recorded at 4 distances (1m, 5m, 10m, 15m, column 'distance'). The data includes ambient (background) noise selections for each distances. The data was created by the function \code{\link[warbleR]{selection_table}} from the warbleR package.
}
\keyword{datasets}
