% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playback_est_unaligned-data.R
\docType{data}
\name{playback_est_unaligned}
\alias{playback_est_unaligned}
\title{Extended selection table with re-recorded playbacks before alignment}
\format{Extended selection table object in the \code{\link{warbleR}} format, which contains annotations and acoustic data}
\source{
Marcelo Araya-Salas
}
\usage{
data(playback_est_unaligned)
}
\description{
The data contains a subet of the selections in the example data 'playback_est' but these selections are not aligned in time with the corresponding reference signals (see \code{\link{spcc_align}} for more details on aligning signals). This data set is intended mostly for using as an example in \code{\link{spcc_align}}. The data contains recordings of \emph{Phaethornis longirostris} (Long-billed Hermit) songs from different song types (column 'signal.type') that were broadcast and re-recorded at 4 distances (1m, 5m, 10m, 15m, column 'distance'). The data was created by the function \code{\link[warbleR]{selection_table}} from the \code{\link{warbleR}} package.
}
\keyword{datasets}
