% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aligned_sounds.R
\name{plot_aligned_sounds}
\alias{plot_aligned_sounds}
\title{Plot spectrograms to check test sound files alignment}
\usage{
plot_aligned_sounds(
  X,
  hop.size = getOption("hop.size", 11.6),
  wl = getOption("wl", NULL),
  ovlp = getOption("ovlp", 50),
  path = getOption("sound.files.path", "."),
  cores = getOption("mc.cores", 1),
  pb = getOption("pb", TRUE),
  collevels = seq(-120, 0, 5),
  palette = viridis::viridis,
  duration = 2,
  mar = 0.2,
  dest.path = getOption("dest.path", "."),
  flim = NULL,
  col = "white",
  width = 7,
  height = 4,
  res = 100,
  label = TRUE,
  fast.spec = FALSE,
  srt = 0,
  cex = 1,
  ...
)
}
\arguments{
\item{X}{Object of class \code{data.frame}, \code{selection_table}, or
\code{extended_selection_table} (the last 2 classes are created by
\code{\link[warbleR:selection_table]{warbleR::selection_table()}} from the \strong{warbleR} package) with
the test sound files' annotations. Must contain the following
columns: 1) \code{sound.files}: name of the \code{.wav} files, 2) \code{selec}:
unique selection identifier (within a sound file), 3) \code{start}:
start time and 4) \code{end}: end time of selections, 5)
\code{bottom.freq}: low frequency for bandpass, 6) \code{top.freq}: high
frequency for bandpass, and 7) \code{sound.id}: ID of sounds used to
identify counterparts across distances. Each sound must have a
unique ID within a distance.}

\item{hop.size}{Numeric vector of length 1 specifying the time window
duration (in ms). Default \code{11.6} ms, which is equivalent to 512
\code{wl} for a 44.1 kHz sampling rate. Ignored if \code{wl} is supplied. Can
be set globally for the current R session via the \code{"hop.size"}
option (see \code{\link[=options]{options()}}).}

\item{wl}{Numeric vector of length 1 specifying the window length of
the spectrogram. Default \code{NULL}. Ignored if \code{bp = NULL}. If
supplied, \code{hop.size} is ignored.}

\item{ovlp}{Numeric vector of length 1 specifying the percentage of
overlap between two consecutive windows, as in
\code{\link[seewave:spectro]{seewave::spectro()}}. Default \code{0}. Can be set globally for the
current R session via the \code{"ovlp"} option (see \code{\link[=options]{options()}}).}

\item{path}{Character string containing the directory path where the
sound files are found. Only needed when \code{X} is not an extended
selection table. If not supplied the current working directory is
used. Can be set globally for the current R session via the
\code{"sound.files.path"} option (see \code{\link[=options]{options()}}).}

\item{cores}{Numeric vector of length 1. Controls whether parallel
computing is applied by specifying the number of cores to be used.
Default \code{1} (i.e. no parallel computing). Can be set globally for
the current R session via the \code{"mc.cores"} option (see
\code{\link[=options]{options()}}).}

\item{pb}{Logical argument to control if progress bar is shown.
Default \code{TRUE}. Can be set globally for the current R session via
the \code{"pb"} option (see \code{\link[=options]{options()}}).}

\item{collevels}{Numeric vector of length 3. Specifies levels to
partition the amplitude range of the spectrogram (in dB). The
more levels, the higher the resolution of the spectrogram. Default
\code{seq(-40, 0, 1)}. \code{seq(-115, 0, 1)} will produce spectrograms
similar to other acoustic analysis software packages.}

\item{palette}{Color palette function for the spectrogram. Default
\code{\link[viridis:reexports]{viridis::viridis()}}. See \code{\link[seewave:spectro]{seewave::spectro()}} for more palettes.
Palettes such as \code{monitoR::gray.2} may work better when
\code{fast.spec = TRUE}.}

\item{duration}{Numeric vector of length 1. Specifies the overall
duration of the clip that will be plotted. Notice that only the
initial part of the test files is plotted, as this is usually
enough to tell the precision of the alignment.}

\item{mar}{Numeric vector of length 1. Specifies the margins
adjacent to the start of the first annotation to be included in
the plot.}

\item{dest.path}{Character string containing the directory path
where the image files will be saved. If not supplied the current
working directory will be used instead. Can be set globally for
the current R session via the \code{"dest.path"} option (see
\code{\link[=options]{options()}}).}

\item{flim}{Numeric vector of length 2 indicating the highest and
lowest frequency limits (kHz) of the spectrogram, as in
\code{\link[seewave:spectro]{seewave::spectro()}}. Default \code{NULL}, which will plot spectrograms
in the full frequency range (0 - nyquist frequency).}

\item{col}{Character string controlling the color of lines and
sound ID labels.}

\item{width}{Numeric vector of length 1. Single value (in inches)
indicating the width of the output image files. Default \code{7}.}

\item{height}{Numeric vector of length 1. Single value (in inches)
indicating the height of the output image files. Default \code{4}.}

\item{res}{Numeric argument of length 1. Controls image resolution.
Default \code{100} (faster), although 300-400 is recommended for
publication/presentation quality.}

\item{label}{Logical to control if labels (from the \code{sound.id}
column in \code{X}) are plotted. Default \code{TRUE}.}

\item{fast.spec}{Logical. If \code{TRUE}, the \code{image} function is used
internally to create spectrograms, which substantially increases
performance (much faster), although some options become
unavailable, such as \code{collevels} (amplitude scale). Default
\code{FALSE}.}

\item{srt}{Numeric argument of length 1. The rotation (in degrees)
of the sound ID labels. Default \code{0}.}

\item{cex}{Numeric argument of length 1 controlling the size of
sound ID text labels. Default \code{1}.}

\item{...}{Additional arguments to be passed to the internal
spectrogram-creating function for customizing graphical output.
The function is a modified version of \code{\link[seewave:spectro]{seewave::spectro()}}, so it
takes the same arguments.}
}
\value{
Image files in \code{jpeg} format with spectrograms in the working
directory, one for each sound file in \code{X}. It also returns the file
path of the images invisibly.
}
\description{
\code{plot_aligned_sounds()} plots spectrograms to visually inspect
alignment precision on test sound files.
}
\details{
This function aims to simplify the evaluation of the alignment of
test sound files from \code{\link[=align_test_files]{align_test_files()}}. The function creates a
single spectrogram for each sound file (saved at \code{dest.path}).
Spectrograms include the first few seconds of the sound files
(controlled by \code{duration}), which is usually enough to tell the
precision of the alignment. The plots include vertical lines
denoting the start and end of each sound, as well as the sound ID
(\code{sound.id} column in \code{X}). Note that no plot is created in the R
graphic device.
}
\examples{
{
  # load example data
  data("test_sounds_est")

  # plot (look into temporary working directory `tempdir()`)
  plot_aligned_sounds(X = test_sounds_est, dest.path = tempdir(), duration = 3, ovlp = 0)
}

}
\references{
Araya-Salas, M., Grabarczyk, E. E., Quiroz-Oliva, M., Garcia-Rodriguez, A., & Rico-Guevara, A. (2025). Quantifying degradation in animal acoustic signals with the R package baRulho. Methods in Ecology and Evolution, 00, 1-12. https://doi.org/10.1111/2041-210X.14481
}
\seealso{
\code{\link[=manual_realign]{manual_realign()}} and \code{\link[=auto_realign]{auto_realign()}}, for fixing
misalignments; \code{\link[=find_markers]{find_markers()}} and \code{\link[=align_test_files]{align_test_files()}}, used
upstream to produce the input for this function.

Other test sound alignment: 
\code{\link{align_test_files}()},
\code{\link{auto_realign}()},
\code{\link{find_markers}()},
\code{\link{manual_realign}()}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
\concept{test sound alignment}
