% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers.R
\name{search_templates}
\alias{search_templates}
\title{alternative name for \code{\link{find_markers}}}
\usage{
search_templates(
  X,
  template.rows = NULL,
  markers = c("start_marker", "end_marker"),
  test.files = NULL,
  path = NULL,
  pb = TRUE,
  cores = 1,
  ...
)
}
\description{
alternative name for \code{\link{find_markers}}
}
\details{
see \code{\link{find_markers}} for documentation. \code{\link{search_templates}} will be deprecated in future versions.
}
\keyword{internal}
