% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_to_noise_ratio.R
\name{snr}
\alias{snr}
\title{alternative name for \code{\link{signal_to_noise_ratio}}}
\usage{
snr(
  X,
  mar,
  parallel = 1,
  pb = TRUE,
  eq.dur = FALSE,
  noise.ref = "adjacent",
  type = 1,
  bp = "freq.range",
  output = "est",
  hop.size = 1,
  wl = NULL
)
}
\description{
alternative name for \code{\link{signal_to_noise_ratio}}
}
\details{
see \code{\link{signal_to_noise_ratio}} for documentation. \code{\link{snr}} will be deprecated in future versions.
}
\keyword{internal}
