% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realign_test_sounds.R
\name{spcc_align}
\alias{spcc_align}
\title{alternative name for \code{\link{auto_realign}}}
\usage{
spcc_align(
  X,
  parallel = NULL,
  cores = getOption("mc.cores", 1),
  pb = getOption("pb", TRUE),
  hop.size = getOption("hop.size", 11.6),
  wl = getOption("wl", NULL),
  ovlp = getOption("ovlp", 90),
  wn = "hanning"
)
}
\description{
alternative name for \code{\link{auto_realign}}
}
\details{
see \code{\link{auto_realign}} for documentation. \code{\link{spcc_align}} will be deprecated in future versions.
}
\keyword{internal}
