% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcc.R
\name{spcc_distortion}
\alias{spcc_distortion}
\title{alternative name for \code{\link{spcc}}}
\usage{
spcc_distortion(
  X,
  parallel = NULL,
  cores = getOption("mc.cores", 1),
  pb = getOption("pb", TRUE),
  method = getOption("method", 1),
  cor.method = "pearson",
  output = NULL,
  hop.size = getOption("hop.size", 11.6),
  wl = getOption("wl", NULL),
  ovlp = getOption("ovlp", 90),
  wn = "hanning",
  path = getOption("sound.files.path", ".")
)
}
\description{
alternative name for \code{\link{spcc}}
}
\details{
see \code{\link{spcc}} for documentation. \code{\link{spcc_distortion}} will be deprecated in future versions.
}
\keyword{internal}
