% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_correlation.R
\name{spectrum_correlation}
\alias{spectrum_correlation}
\title{Measure frequency spectrum correlation}
\usage{
spectrum_correlation(
  X,
  cores = getOption("mc.cores", 1),
  pb = getOption("pb", TRUE),
  cor.method = c("pearson", "spearman", "kendall"),
  spec.smooth = getOption("spec.smooth", 5),
  hop.size = getOption("hop.size", 11.6),
  wl = getOption("wl", NULL),
  ovlp = getOption("ovlp", 70),
  path = getOption("sound.files.path", "."),
  n.bins = 100
)
}
\arguments{
\item{X}{The output of \code{\link[=set_reference_sounds]{set_reference_sounds()}}, an object of class
\code{data.frame}, \code{selection_table}, or \code{extended_selection_table} (the
last 2 classes are created by \code{\link[warbleR:selection_table]{warbleR::selection_table()}} from the
\strong{warbleR} package) with the test sound files' annotations. Must
contain the following columns: 1) \code{sound.files}: name of the \code{.wav}
files, 2) \code{selec}: unique selection identifier (within a sound
file), 3) \code{start}: start time and 4) \code{end}: end time of selections,
5) \code{bottom.freq}: low frequency for bandpass, 6) \code{top.freq}: high
frequency for bandpass, 7) \code{sound.id}: ID of sounds used to
identify counterparts across distances, and 8) \code{reference}:
identity of sounds to be used as reference for each test sound
(row). See \code{\link[=set_reference_sounds]{set_reference_sounds()}} for more details on the
structure of \code{X}.}

\item{cores}{Numeric vector of length 1. Controls whether parallel
computing is applied by specifying the number of cores to be used.
Default \code{1} (i.e. no parallel computing). Can be set globally for
the current R session via the \code{"mc.cores"} option (see
\code{\link[=options]{options()}}).}

\item{pb}{Logical argument to control if progress bar is shown.
Default \code{TRUE}. Can be set globally for the current R session via
the \code{"pb"} option (see \code{\link[=options]{options()}}).}

\item{cor.method}{Character string indicating the correlation
coefficient to be applied (\code{"pearson"}, \code{"spearman"}, or
\code{"kendall"}, see \code{\link[stats:cor]{stats::cor()}}).}

\item{spec.smooth}{Numeric vector of length 1 determining the
length of the sliding window used for a sum smooth for power
spectrum calculation (in kHz). Default \code{5}.}

\item{hop.size}{Numeric vector of length 1 specifying the time window
duration (in ms). Default \code{11.6} ms, which is equivalent to 512
\code{wl} for a 44.1 kHz sampling rate. Ignored if \code{wl} is supplied. Can
be set globally for the current R session via the \code{"hop.size"}
option (see \code{\link[=options]{options()}}).}

\item{wl}{A vector with a single even integer number specifying the
window length of the spectrogram. Default \code{NULL}. If supplied,
\code{hop.size} is ignored. Odd integers will be rounded up to the
nearest even number. Can be set globally for the current R session
via the \code{"wl"} option (see \code{\link[=options]{options()}}).}

\item{ovlp}{Numeric vector of length 1 specifying the percentage of
overlap between two consecutive windows, as in
\code{\link[seewave:spectro]{seewave::spectro()}}. Default \code{70}. Can be set globally for the
current R session via the \code{"ovlp"} option (see \code{\link[=options]{options()}}).}

\item{path}{Character string containing the directory path where the
sound files are found. Only needed when \code{X} is not an extended
selection table. If not supplied the current working directory is
used. Can be set globally for the current R session via the
\code{"sound.files.path"} option (see \code{\link[=options]{options()}}).}

\item{n.bins}{Numeric vector of length 1 specifying the number of
frequency bins to use for representing power spectra. Default
\code{100}. If \code{NULL}, the raw power spectrum is used (note that this
can result in high RAM memory usage for large data sets). Power
spectrum values are interpolated using \code{\link[stats:approxfun]{stats::approx()}}.}
}
\value{
Object \code{X} with an additional column, \code{spectrum.correlation},
containing the computed frequency spectrum correlation
coefficients.
}
\description{
\code{spectrum_correlation()} measures frequency spectrum correlation of
sounds referenced in an extended selection table. Spectral
correlation measures the similarity of two sounds in the frequency
domain.
}
\details{
The function measures the spectral correlation coefficients of
sounds in which a reference playback has been re-recorded at
increasing distances. Values range from 1 (identical frequency
spectrum, i.e. no degradation) to 0. The \code{sound.id} column must be
used to tell the function to only compare sounds belonging to the
same category (e.g. song-types). The function will then compare
each sound to the corresponding reference sound. Two methods for
computing spectral correlation are provided (see the \code{method}
argument). The function uses \code{\link[seewave:meanspec]{seewave::meanspec()}} internally to
compute power spectra. Use \code{\link[=spectrum_blur_ratio]{spectrum_blur_ratio()}} to extract raw
spectra values. \code{NA} is returned if at least one of the power
spectra cannot be computed.
}
\examples{
{
  # load example data
  data("test_sounds_est")

  # method 1
  # add reference column
  Y <- set_reference_sounds(X = test_sounds_est)

  # run spectrum correlation
  spectrum_correlation(X = Y)

  # method 2
  Y <- set_reference_sounds(X = test_sounds_est, method = 2)
  # spectrum_correlation(X = Y)
}

}
\references{
Araya-Salas, M., Grabarczyk, E. E., Quiroz-Oliva, M., Garcia-Rodriguez, A., & Rico-Guevara, A. (2025). Quantifying degradation in animal acoustic signals with the R package baRulho. Methods in Ecology and Evolution, 00, 1-12. https://doi.org/10.1111/2041-210X.14481
Apol, C.A., Sturdy, C.B. & Proppe, D.S. (2017). Seasonal variability in habitat structure may have shaped acoustic signals and repertoires in the black-capped and boreal chickadees. Evol Ecol. 32:57-74.
}
\seealso{
\code{\link[=envelope_correlation]{envelope_correlation()}} and \code{\link[=spectrum_blur_ratio]{spectrum_blur_ratio()}}.

Other quantify degradation: 
\code{\link{blur_ratio}()},
\code{\link{detection_distance}()},
\code{\link{envelope_correlation}()},
\code{\link{plot_blur_ratio}()},
\code{\link{plot_degradation}()},
\code{\link{set_reference_sounds}()},
\code{\link{signal_to_noise_ratio}()},
\code{\link{spcc}()},
\code{\link{spectrum_blur_ratio}()},
\code{\link{tail_to_signal_ratio}()}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
\concept{quantify degradation}
