% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmospheric_attenuation.R
\name{atmospheric_attenuation}
\alias{atmospheric_attenuation}
\title{Estimate atmospheric attenuation}
\usage{
atmospheric_attenuation(f, temp, RH, p = 101325, formula = NULL, spi = NULL, dist)
}
\arguments{
\item{f}{numeric vector of length 1 with frequency (in Hertz).}

\item{temp}{numeric vector of length 1 with frequency (in Celsius).}

\item{RH}{numeric vector of length 1 with relative humidity}

\item{p}{numeric vector of length 1 with ambient pressure in Pa (standard: 101325, default).}

\item{formula}{DEPRECATED.}

\item{spi}{numeric vector of length 1 with the initial sound pressure in Pa. Default is \code{NULL}. DEPRECATED.}

\item{dist}{numeric vector of length 1 with distance (m) over which a sound propagates.}
}
\value{
Returns atmospheric attenuation (in dB) of sound based on supplied parameters.
}
\description{
\code{atmospheric_attenuation} estimates atmospheric attenuation and atmospheric absorption.
}
\details{
Calculate the atmospheric attenuation based on temperature, relative humidity, pressure and sound frequency.
}
\examples{
{
# measure atmospheric attenuation formula 1
atmospheric_attenuation(f = 20000, temp = 20, RH = 90, p = 88000, dist = 30)
}

}
\references{
{
Araya-Salas, M. (2020). baRulho: baRulho: quantifying habitat-induced degradation of (animal) acoustic signals in R. R package version 1.0.2
}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
