% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attenuation.R
\name{attenuation}
\alias{attenuation}
\title{Estimate attenuation of sound pressure level}
\usage{
attenuation(f, temp = 20, RH = 60, pa = 101325, dist0, dist, att.coef = 0.02)
}
\arguments{
\item{f}{numeric vector of length 1 with frequency (in Hertz).}

\item{temp}{numeric vector of length 1 with frequency (in Celsius). Default is 20.}

\item{RH}{numeric vector of length 1 with relative humidity (in percentage). Default is 60.}

\item{pa}{numeric vector of length 1 with ambient pressure in Pa (standard: 101325, default).}

\item{dist0}{numeric vector of length 1 with distance (m) for the reference SPL.}

\item{dist}{numeric vector of length 1 with distance (m) over which a sound propagates.}

\item{att.coef}{attenuation coefficient of the habitat (in dB/kHz/m).}
}
\value{
Returns the geometric, atmospheric and habitat attenuation (in dB) as well as the combined attenuation.
}
\description{
\code{attenuation} estimates atmospheric attenuation and atmospheric absorption.
}
\details{
Calculate the atmospheric attenuation based on temperature, relative humidity, pressure and sound frequency.
}
\examples{
{
# measure attenuation
attenuation(f = 2000, dist = 50, dist0 = 1)
}

}
\references{
{
Araya-Salas, M. (2020). baRulho: baRulho: quantifying habitat-induced degradation of (animal) acoustic signals in R. R package version 1.0.2
}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
