% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_realign.R
\name{auto_realign}
\alias{auto_realign}
\title{Fix misaligned start and end of test sounds}
\usage{
auto_realign(X, parallel = NULL, cores = getOption("mc.cores", 1),
pb = getOption("pb", TRUE), hop.size = getOption("hop.size", 11.6),
wl = getOption("wl", NULL), ovlp = getOption("ovlp", 90), wn = 'hanning')
}
\arguments{
\item{X}{object of class 'extended_selection_table' created by the function \code{\link[warbleR]{selection_table}} from the warbleR package. The object must include the following additional columns: 'sound.id', 'bottom.freq' and 'top.freq'.}

\item{parallel}{DEPRECATED. Use 'cores' instead.}

\item{cores}{Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{pb}{Logical argument to control if progress bar is shown. Default is \code{TRUE}.}

\item{hop.size}{A numeric vector of length 1 specifying the time window duration (in ms). Default is 11.6 ms, which is equivalent to 512 wl for a 44.1 kHz sampling rate. Ignored if 'wl' is supplied.}

\item{wl}{A numeric vector of length 1 specifying the window length of the spectrogram, default
is NULL. If supplied, 'hop.size' is ignored.}

\item{ovlp}{Numeric vector of length 1 specifying \% of overlap between two
consecutive windows, as in \code{\link[seewave]{spectro}}. Default is 90. High values of ovlp
slow down the function but produce more accurate results.}

\item{wn}{A character vector of length 1 specifying the window name as in \code{\link[seewave]{ftwindow}}.}
}
\value{
Object 'X' in which time parameters (columns 'start' and 'end') have been tailored to more closely match the start and end of the reference sound.
}
\description{
\code{auto_realign} fixes misaligned start and end of test sounds in an extended selection table using spectrographic cross-correlation
}
\details{
This function uses spectrographic cross-correlation to align the position in time of sounds with regard to a reference sound. The sound recorded at the closest distance to the source is used as reference. Precise alignment is crucial for downstream measures of sound degradation. The function calls warbleR's \code{\link[warbleR]{cross_correlation}} internally to align sounds using cross-correlation. The output extended selection table contains the new start and end values after alignment.
}
\examples{
{
  # load example data
  data("test_sounds_est")

  # create "unaligned_test_sounds_est" by
  # adding noise to "test_sounds_est" start and end
  unaligned_test_sounds_est <- test_sounds_est
  set.seed(123)
  noise_time <- sample(c(0.005, -0.005, 0.006, -0.006, 0, 0.002, -0.002),
    nrow(unaligned_test_sounds_est),
    replace = TRUE
  )

  attr(unaligned_test_sounds_est, "check.res")$start <-
    unaligned_test_sounds_est$start <- unaligned_test_sounds_est$start + noise_time

  attr(unaligned_test_sounds_est, "check.res")$end <- unaligned_test_sounds_est$end <-
    unaligned_test_sounds_est$end + noise_time

  # re align

  auto_realign(X = unaligned_test_sounds_est)
}

}
\references{
{
Araya-Salas, M. (2020). baRulho: baRulho: quantifying degradation of (animal) acoustic signals in R. R package version 1.0.2

Clark, C.W., Marler, P. & Beeman K. (1987). Quantitative analysis of animal vocal phonology: an application to Swamp Sparrow song. Ethology. 76:101-115.
}
}
\seealso{
\code{\link{blur_ratio}}, \code{\link[warbleR]{cross_correlation}}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
