% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baRulho-package.R
\docType{package}
\name{baRulho-package}
\alias{baRulho}
\alias{baRulho-package}
\title{baRulho: Quantifying (Animal) Sound Degradation}
\description{
Intended to facilitate acoustic analysis of (animal) sound transmission experiments, which typically aim to quantify changes in signal structure when transmitted in a given habitat by broadcasting and re-recording animal sounds at increasing distances. The package offers a workflow with functions to prepare the data set for analysis as well as to calculate and visualize several degradation metrics, including blur ratio, signal-to-noise ratio, excess attenuation and envelope correlation among others (Dabelsteen et al 1993 \doi{10.1121/1.406682}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maRce10/baRulho}
  \item Report bugs at \url{https://github.com/maRce10/baRulho/issues}
}

}
\author{
\strong{Maintainer}: Marcelo Araya-Salas \email{marcelo.araya@ucr.ac.cr}

}
\keyword{internal}
