% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baRulho-package.R
\docType{package}
\name{baRulho}
\alias{baRulho}
\title{baRulho: quantifying habitat-induced acoustic signal degradation}
\description{
`baRulho` is a package intended to quantify habitat-induced degradation of (animal) acoustic signals.
}
\details{
The main features of the package are:
  \itemize{
  \item Loops to apply tasks through acoustic signals referenced in an extended selection table
  \item The comparison of playback signals re-recorded at different distances 
  }
Most functions allow the parallelization of tasks, which distributes the tasks among several processors to improve computational efficiency.

License: GPL (>= 2)
}
\author{
Marcelo Araya-Salas
  
  Maintainer: Marcelo Araya-Salas (\email{marceloa27@gmail.com})
}
