% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur_ratio.R
\name{blur_ratio}
\alias{blur_ratio}
\title{Measure blur ratio}
\usage{
blur_ratio(X, parallel = 1, pb = TRUE, method = 1,  
 ssmooth = NULL, msmooth = NULL, output = "est")
}
\arguments{
\item{X}{object of class 'selection_table', 'extended_selection_table' created by the function \code{\link[warbleR]{selection_table}} from the warbleR package.}

\item{parallel}{Numeric. Controls whether parallel computing is applied.
It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).
If \code{NULL} (default) then the current working directory is used.}

\item{pb}{Logical argument to control if progress bar is shown. Default is \code{TRUE}. It can also be
set globally using the 'pb' option (see \code{\link{warbleR_options}}).}

\item{method}{Numeric vector of length 1 to indicate the 'experimental design' for measuring envelope correlation. Two methods are available:
\itemize{
\item \code{1}: compare all signals with their counterpart that was recorded at the closest distance to source (e.g. compare a signal recorded at 5m, 10m and 15m with its counterpart recorded at 1m). This is the default method. 
\item \code{2}: compare all signals with their counterpart recorded at the distance immediately before (e.g. a signal recorded at 10m compared with the same signal recorded at 5m, then signal recorded at 15m compared with same signal recorded at 10m and so on).
}}

\item{ssmooth}{Numeric vector of length 1 determining the length of the sliding window used for a sum smooth for amplitude envelope calculation (used internally by \code{\link[seewave]{env}}).}

\item{msmooth}{Numeric vector of length 2 to smooth the amplitude envelope with a mean sliding window for amplitude envelope calculation. The first element is the window length (in number of amplitude values) and the second one the window overlap (used internally by \code{\link[seewave]{env}}).}

\item{output}{Character vector of length 1 to determine if an extended selection table ('est') or a data frame ('data.frame') is returned.}
}
\value{
Data frame similar to input data, but also includes a new column (blur.ratio)
with the excess attenuation values.
}
\description{
\code{blur_ratio} measures blur ratio in signals referenced in a extended selection table.
}
\details{
Blur ratio measures the degradation of sound as a function of the change in signal energy in the time domain as described by Dabelsteen et al 1993. The goal of the function is to measure the blur ratio on signals in which a master playback has been re-recorded at different distances. The 'signal.id' column must be used to tell the function to only compare signals belonging to the same category (e.g. song-types). Two methods for calculating excess attenuation are provided
}
\examples{
{
# load example data
data("playback_est")

# remove noise selections
playback_est <- playback_est[playback_est$signal.id != "noise", ]

# using method 1
blur_ratio(X = playback_est)

# using method 2
blur_ratio(X = playback_est, method = 2)
}

}
\author{
Marcelo Araya-Salas (\email{marceloa27@gmail.com}) #' @references {
Dabelsteen, T., Larsen, O. N., & Pedersen, S. B. (1993). Habitat-induced degradation of sound signals: Quantifying the effects of communication sounds and bird location on blur ratio, excess attenuation, and signal-to-noise ratio in blackbird song. The Journal of the Acoustical Society of America, 93(4), 2206.

Araya-Salas, M. (2019), baRulho: a R package to evaluate habitat-induced degradation of (animal) acoustic signals. R package version 1.0.0
}
}
