% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_distance.R
\name{detection_distance}
\alias{detection_distance}
\title{Measure spectrographic cross-correlation as a measure of sound distortion}
\usage{
detection_distance(X, cores = 1, pb = TRUE, output = "est",
hop.size = 11.6, wl = NULL, path = NULL, spl = NULL, spl.cutoff, temp = 20, 
rh = 60, pa = 101325, hab.att.coef = 0.02, max.distance = 1000,
 resolution = 0.1, subtract.bgn = TRUE, envelope = "abs", mar = NULL)
}
\arguments{
\item{X}{Object of class 'data.frame', 'selection_table' or 'extended_selection_table' (the last 2 classes are created by the function \code{\link[warbleR]{selection_table}} from the warbleR package) with the reference to the sounds in the master sound file. Must contain the following columns: 1) "sound.files": name of the .wav files, 2) "selec": unique selection identifier (within a sound file), 3) "start": start time and 4) "end": end time of selections, 5)  "bottom.freq": low frequency for bandpass, 6) "top.freq": high frequency for bandpass and 7) "sound.id": ID of sounds used to identify counterparts across distances. Each sound must have a unique ID within a distance.}

\item{cores}{Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{pb}{Logical argument to control if progress bar is shown. Default is \code{TRUE}.}

\item{output}{Character vector of length 1 to determine if an extended selection table ('est', default) or a data frame ('data.frame').}

\item{hop.size}{A numeric vector of length 1 specifying the time window duration (in ms). Default is 11.6 ms, which is equivalent to 512 wl for a 44.1 kHz sampling rate. Ignored if 'wl' is supplied.}

\item{wl}{A numeric vector of length 1 specifying the window length of the spectrogram, default
is \code{NULL}. If supplied, 'hop.size' is ignored.}

\item{path}{Character string containing the directory path where the sound files are found. Only needed when 'X' is not an extended selection table.}

\item{spl}{A numeric vector of length 1 specifying the sound pressure level of sounds. If not supplied then it will be measured from the sounds  themselves.}

\item{spl.cutoff}{A numeric vector of length 1 specifying the sound pressure level cutoff to define if the sound is no longer detected. Ideally it should be estimated based on the sound detection threshold of the species.}

\item{temp}{Numeric vector of length 1 with frequency (in Celsius). Default is 20.}

\item{rh}{Numeric vector of length 1 with relative humidity (in percentage). Default is 60.}

\item{pa}{Numeric vector of length 1 with ambient pressure in Pa (standard: 101325, default). Used for Atmospheric attenuation.}

\item{hab.att.coef}{Attenuation coefficient of the habitat (in dB/kHz/m).}

\item{max.distance}{Numeric vector of length 1 with the maximum distance (in m) at which detection would be evaluated. Note that the function calculates the expected sound pressure level values along a vector of distances to find the distance at which the expected sound pressure level equates 'spl.cutoff'. Default is 1000 m.}

\item{resolution}{NUmeric vector of length 1 with the distance resolution (in m) for estimated detection distance. Higher resolutions take longer to estimate. Default is 0.1 m.}

\item{subtract.bgn}{Logical argument to control if SPL from background noise is excluded from the measured signal SPL. Default is \code{FALSE}.}

\item{envelope}{Character string vector with the method to calculate amplitude envelopes (in which SPL is measured, used required if 'spl' is not supplied), as in \code{\link[seewave]{env}}. Must be either 'abs' (absolute envelope, default) or 'hil' (Hilbert transformation).}

\item{mar}{numeric vector of length 1. Specifies the margins adjacent to
the start and end points of selection over which to measure background noise. This is required to subtract background noise sound pressure level (so only needed when 'subtract.bgn = TRUE').}
}
\value{
Returns the geometric, atmospheric and habitat attenuation (in dB) as well as the combined attenuation.

A data frame, or extended selection table similar to input data (depending on argument 'output'), but including two new columns (reference and detection.distance).
}
\description{
\code{detection_distance} measures spectrographic cross-correlation as a measure of sound distortion in sounds referenced in an extended selection table.
}
\details{
The function computes the maximum distance at which a sound would be detected, which is calculated as the distance in which the sound pressure level (SPL) goes below the specified SPL cutoff ('spl.cutoff')). The function uses internally \code{\link{attenuation}} to estimate SPL at increasing values until it reaches the defined cutoff. The peak frequency (caculated on the power spectrum of the reference sound) of the reference sound for each sound ID is used as the carrier frequency for distance estimation. \strong{This function assumes that all recordings have been made at the same recording volume}.
}
\examples{
{
# load example data
data("playback_est")

detection_distance(X = playback_est[playback_est$distance == 1,], spl.cutoff = 5, mar = 0.05)
}

}
\references{
{
Araya-Salas, M. (2020). baRulho: baRulho: quantifying habitat-induced degradation of (animal) acoustic signals in R. R package version 1.0.2

Clark, C.W., Marler, P. & Beeman K. (1987). Quantitative analysis of animal vocal phonology: an application to Swamp Sparrow song. Ethology. 76:101-115.
}
}
\seealso{
\code{\link{attenuation}}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
