% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_attenuation.R
\name{excess_attenuation}
\alias{excess_attenuation}
\title{Measure excess attenuation}
\usage{
excess_attenuation(X, parallel = 1, pb = TRUE, method = 1, 
bp = NULL, wl = 10, output = "est")
}
\arguments{
\item{X}{object of class 'selection_table', 'extended_selection_table' created by the function \code{\link[warbleR]{selection_table}} from the warbleR package.}

\item{parallel}{Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{pb}{Logical argument to control if progress bar is shown. Default is \code{TRUE}.}

\item{method}{Numeric vector of length 1 to indicate the 'experimental design' for measuring excess attenuation. Two methods are available:
\itemize{
\item \code{1}: compare all signals with their counterpart that was recorded at the closest distance to source (e.g. compare a signal recorded at 5m, 10m and 15m with its counterpart recorded at 1m). This is the default method. 
\item \code{2}: compare all signals with their counterpart recorded at the distance immediately before (e.g. a signal recorded at 10m compared with the same signal recorded at 5m, then signal recorded at 15m compared with same signal recorded at 10m and so on).
}}

\item{bp}{Numeric vector of length 2 giving the lower and upper limits of a frequency bandpass filter (in kHz). Default is \code{NULL}.}

\item{wl}{A numeric vector of length 1 specifying the window length of the spectrogram for applying bandpass. Default 
 is 10. Ignored if \code{bp = NULL}.
Note that lower values will increase time resolution, which is more important for amplitude ratio calculations.}

\item{output}{Character vector of length 1 to determine if an extended selection table ('est', default) or a data frame ('data.frame') is returned.}
}
\value{
Data frame or extended selection table (depending on 'output' argument) similar to input data, but also includes a new column (excess.attenuation)
with the excess attenuation values.
}
\description{
\code{excess_attenuation} measures excess attenuation in signals referenced in an extended selection table.
}
\details{
Excess attenuation is the amplitude loss of a sound in excess of that due to spherical spreading. With every doubling of distance, sounds attenuate with a 6 dB loss of amplitude (Morton, 1975; Marten & Marler, 1977). Any additional loss of amplitude results in excess attenuation, or energy loss in excess of that expected to occur with distance via spherical spreading due to atmospheric conditions or habitat (Wiley & Richards, 1978).

The goal of the function is to measure the excess attenuation on signals in which a reference playback has been re-recorded at increasing distances. The 'signal.id' column must be used to indicate which signals belonging to the same category (e.g. song-types). The function will then compare each signal type to the corresponding reference signal. Two methods for calculating excess attenuation are provided (see 'method' argument).
}
\examples{
{
# load example data
data("playback_est")

# remove noise selections
playback_est <- playback_est[playback_est$signal.id != "noise", ]

# using method 1
excess_attenuation(X = playback_est)

# using method 2
excess_attenuation(X = playback_est, method = 2)
}

}
\author{
Marcelo Araya-Salas (\email{marceloa27@gmail.com}) #' @references {
Araya-Salas, M. (2019), baRulho: a R package to quantify habitat-induced degradation of (animal) acoustic signals. R package version 1.0.0

Marten, K., & Marler, P. (1977). Sound transmission and its significance for animal vocalization. Behavioral ecology and sociobiology, 2(3), 271-290.

Morton, E. S. (1975). Ecological sources of selection on avian sounds. The American Naturalist, 109(965), 17-34.
}
}
