% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snr_attenuation.R
\name{snr_attenuation}
\alias{snr_attenuation}
\title{Measure attenuation as signal-to-noise ratio}
\usage{
snr_attenuation(X, mar, parallel = 1, pb = TRUE, eq.dur = FALSE,
noise.ref = "adjacent", type = 1, bp = NULL, wl = 10, output = "est")
}
\arguments{
\item{X}{object of class 'selection_table', 'extended_selection_table' created by the function \code{\link[warbleR]{selection_table}} from the warbleR package.}

\item{mar}{numeric vector of length 1. Specifies the margins adjacent to
the start and end points of selection over which to measure noise.}

\item{parallel}{Numeric. Controls whether parallel computing is applied.
It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).
If \code{NULL} (default) then the current working directory is used.}

\item{pb}{Logical argument to control if progress bar is shown. Default is \code{TRUE}. It can also be
set globally using the 'pb' option (see \code{\link{warbleR_options}}).}

\item{eq.dur}{Logical. Controls whether the noise segment that is measured has the same duration 
than the signal (if \code{TRUE}. Default is \code{FALSE}). If \code{TRUE} then 'mar' and 'noise.ref' arguments are ignored.}

\item{noise.ref}{Character vector of length 1 to determined if a noise segment to be used for measuring ambient noise. Two options are available: 
\itemize{
\item \code{adjacent}: measure ambient noise right before the signal (using argument 'mar' to define duration of noise segments). If several 'noise' selections by sound file are supplied, then the root mean square of the amplitude envelope will be averaged across those selections.
\item \code{custom}: measure noise segments referenced in the selection table (labeled as 'noise' in the 'signal.id' column). Those segments will be used to apply the same noise reference to all signals in a sound file. Therefore, at least one 'noise' selection for each sound file must be provided.
}}

\item{type}{Numeric. Determine the formula to be used to calculate the signal-to-noise ratio (S = signal
, N = background noise): 
\itemize{
\item \code{1}: ratio of S amplitude envelope quadratic mean to N amplitude envelope quadratic mean
 (\code{rms(env(S))/rms(env(N))})
\item \code{2}: ratio of the difference between S amplitude envelope quadratic mean and N amplitude envelope quadratic mean to N amplitude envelope quadratic mean (\code{(rms(env(S)) - rms(env(N)))/rms(env(N))}, as proposed by Dabelsteen et al (1993))
}}

\item{bp}{Numeric vector of length 2 giving the lower and upper limits of a frequency bandpass filter (in kHz). Default is \code{NULL}.}

\item{wl}{A numeric vector of length 1 specifying the window length of the spectrogram for applying bandpass. Default 
 is 10. Ignored if \code{bp = NULL}.
Note that lower values will increase time resolution, which is more important for amplitude ratio calculations.}

\item{output}{Character vector of length 1 to determine if an extended selection table ('est', default) or a data frame ('data.frame') is returned.}
}
\value{
Data frame or extended selection table (depending on 'output' argument) similar to input data, but also includes a new column (snr.attenuation)
with the signal-to-noise ratio values.
}
\description{
\code{snr_attenuation} measures attenuation as signal-to-noise ratio of signals referenced in a extended selection table.
}
\details{
Signal-to-noise ratio is the attenuation is the attenuation of a sound in excess of that due to spherical spreading as described by Dabelsteen et al (1993). The goal of the function is to measure the excess attenuation on signals in which a master playback has been re-recorded at different distances. The 'signal.id' column must be used to indicate which signals belonging to the same category (e.g. song-types). The function will then compared each signal type to its reference. Two methods for calculating excess attenuation are provided (see 'method' argument).
}
\examples{
{
# load example data
data("playback_est")

# using noise reference selections 
snr_attenuation(X = playback_est, mar = 0.05, noise.ref = 'custom')

#' # remove noise selections
playback_est <- playback_est[playback_est$signal.id != "noise", ]
# using margin for noise of 0.05 and adjacent noise reference
snr_attenuation(X = playback_est, mar = 0.05, noise.ref = 'adjacent')
}

}
\author{
Marcelo Araya-Salas (\email{marceloa27@gmail.com}) #' @references {
Dabelsteen, T., Larsen, O. N., & Pedersen, S. B. (1993). Habitat-induced degradation of sound signals: Quantifying the effects of communication sounds and bird location on blur ratio, excess attenuation, and signal-to-noise ratio in blackbird song. The Journal of the Acoustical Society of America, 93(4), 2206.

Araya-Salas, M. (2019), baRulho: a R package to evaluate habitat-induced degradation of (animal) acoustic signals. R package version 1.0.0
}
}
