% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realign_test_sounds.R
\name{spcc_align}
\alias{spcc_align}
\title{alternative name for \code{\link{realign_test_sounds}}}
\usage{
spcc_align(
  X,
  parallel = 1,
  cores = 1,
  pb = TRUE,
  hop.size = 11.6,
  wl = NULL,
  ovlp = 90,
  wn = "hanning"
)
}
\description{
alternative name for \code{\link{realign_test_sounds}}
}
\details{
see \code{\link{realign_test_sounds}} for documentation. \code{\link{spcc_align}} will be deprecated in future versions.
}
\keyword{internal}
