% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcc_distortion.R
\name{spcc_distortion}
\alias{spcc_distortion}
\title{alternative name for \code{\link{spcc}}}
\usage{
spcc_distortion(
  X,
  parallel = 1,
  pb = TRUE,
  method = 1,
  cor.method = "pearson",
  output = "est",
  hop.size = 11.6,
  wl = NULL,
  ovlp = 90,
  wn = "hanning"
)
}
\description{
alternative name for \code{\link{spcc}}
}
\details{
see \code{\link{spcc}} for documentation. \code{\link{spcc_distortion}} will be deprecated in future versions.
}
\keyword{internal}
