% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_correlation.R
\name{spectrum_correlation}
\alias{spectrum_correlation}
\title{Measure frequency spectrum correlation}
\usage{
spectrum_correlation(X, parallel = 1, pb = TRUE, method = 1, cor.method = "pearson", 
output = "est")
}
\arguments{
\item{X}{object of class 'selection_table', 'extended_selection_table' created by the function \code{\link[warbleR]{selection_table}} from the warbleR package.}

\item{parallel}{Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing).
If \code{NULL} (default) then the current working directory is used.}

\item{pb}{Logical argument to control if progress bar is shown. Default is \code{TRUE}.}

\item{method}{Numeric vector of length 1 to indicate the 'experimental design' to measure frequency spectrum correlation. Two methods are available:
\itemize{
\item \code{1}: compare all signals with their counterpart that was recorded at the closest distance to source (e.g. compare a signal recorded at 5m, 10m and 15m with its counterpart recorded at 1m). This is the default method. 
\item \code{2}: compare all signals with their counterpart recorded at the distance immediately before (e.g. a signal recorded at 10m compared with the same signal recorded at 5m, then signal recorded at 15m compared with same signal recorded at 10m and so on).
}}

\item{cor.method}{Character string indicating the correlation coefficient to be applied ("pearson", "spearman", or "kendall", see \code{\link[stats]{cor}}).}

\item{output}{Character vector of length 1 to determine if an extended selection table ('est', default) or a data frame ('data.frame') is returned.}
}
\value{
Data frame or extended selection table (depending on 'output' argument) similar to input data, but also includes a new column ('spectrum.correlation')
with the calculated frequency spectrum correlation coefficients.
}
\description{
\code{spectrum_correlation} measures frequency spectrum correlation of signals referenced in an extended selection table.
}
\details{
frequency spectrum correlation measures the similarity of 2 signals in the time domain. The function measures the spectrum correlation coefficients of signals in which a reference playback has been re-recorded at increasing distances. Values range from 1 (identical frequency spectrum, i.e. no degradation) to 0. The 'signal.id' column must be used to indicate the function to only compare signals belonging to the same category (e.g. song-types). The function will then compare each signal type to the corresponding reference signal. Two methods for calculating spectrum correlation are provided (see 'method' argument).
}
\examples{
{
# load example data
data("playback_est")

# remove noise selections
playback_est <- playback_est[playback_est$signal.id != "noise", ]

# method 1
spectrum_correlation(X = playback_est)

# method 2
spectrum_correlation(X = playback_est, method = 2)
}

}
\references{
{
Araya-Salas, M. (2019), baRulho: a R package to quantify habitat-induced degradation of (animal) acoustic signals. R package version 1.0.0

Apol, C.A., Sturdy, C.B. & Proppe, D.S. (2017). Seasonal variability in habitat structure may have shaped acoustic signals and repertoires in the black-capped and boreal chickadees. Evol Ecol. 32:57-74.
}
}
\author{
Marcelo Araya-Salas (\email{marceloa27@gmail.com})
}
